pyproject.toml
[project]
name = "mmagent"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "pycuda",
    "torch==2.10.0",
    "litellm",
    "python-dotenv",
    "numpy",
]
